Major modification to email archiving.
- Moved processing into the regular execute_analysis function instead of execute_final_analysis.
- Added additional processing to execute_post_analysis to catch URLs and content analyzed later.
- Always updating the email archive database even if the encryption password is not available.
- Update unit tests to reflect these changes.

The issue we were seeing was that since the archive depended on everything else
completing before inserting into the update, if something went wrong then the
archive entry is never written. This fixes that issue by writing the archive
entry right away with the information that is availabe, and then later if more
information is found it is updated.

This was causing issues because, at the time of this commit, remediation
capabilities from the GUI depend on the archives database.

Also, not everyone will be saving a locally encrypted copy of all their emails,
so having that be optional (by setting the saq.ENCRYPTION_PASSWORD to None) is
better.