Added a "dispositioned" analysis mode.
Added a new analysis mode called dispositioned that executes after a user sets
the disposition on an Alert. A very limited set of analysis modules are
configured by default to run in this mode.

Once an analyst sets the disposition of an Alert, analysis should stop in
correlation mode because the analyst has figured it out. However, there may
still be things that need to be done post disposition. This mode fills that
role.