Replaced ACEAlertDispositionAnalyzer with inline engine code.
- start tracking the original analysis mode in the RootAnalysis object
- removed the default_analysis_mode constructor argument
The add_workload function uses the default_analysis_mode out of the configuration file, but the Engine uses it's local copy, which can be passed
in. I removed this functionality to simplify and force everything to just use the value in the configuration file.

- moved the try/catch into the Engine.clear_work_target function (to avoid copy pasta)
- check for Alert disposition while in correlation mode before analysis starts
- replaced ACEAlertDispositionAnalyzer with inline code inside the main engine loop
The ACEAlertDispositionAnalyzer was attempting to do this but it wasn't working. The functionality was moved to the engine itself.

- fixed all the unit testing around this issue