Only retry on network-based errors.
If a submission cannot be sent it can be retried again at a later time. However,
if the failure is because of an unrecoverable error condition (such as the files
getting deleted), it will continue to try it over and over again.

As a TEMPORARY solution this patch only retries on the three network-related
exceptions I've identified so far. The correct solution is to identify the
correct base exception classes and catch those.